requirements.txt
django-phonenumber-field
phonenumbers
Django==2.2.7